// front/src/api/generated/schemas/engineer.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Engineer Route Plan API
 * OpenAPI spec version: 0.1.0
 */
import type { LocalTime } from './localTime';
import type { Point } from './point';
import type { Skill } from './skill';
import type { VehicleType } from './vehicleType';

export interface Engineer {
  /** Идентификатор бригады */
  id: number;
  /** Название бригады */
  name: string;
  vehicle_type: VehicleType;
  /**
     * Навыки бригады, от одного до трёх
     * @minItems 1
     * @maxItems 3
     */
  skills: Skill[];
  shift_start: LocalTime;
  shift_end: LocalTime;
  start: Point;
}